Initial implementation of babel in Jekyll;
This implementation needs a copy of an AMD loader to be on the page in order to
actually work in the browser. I'm thinking that maybe I can load in requirejs or
even almond JS to do this. Another option would be to just write my own AMD loader
which needs to be injected first on the page.

This implementation also has some big limitations. Right now it doesn't handle
directories at all. Also, I haven't tested it while having Jekyll serving files.

I want to take a look at more examples of Jekyll plugins that are doing this kind
of processing (i.e. Sass / Less ) and see if I can learn something about the how
to make this plugin better.

When this plugin is complete, I won't need the Babel gem since it's just a simple
Ruby wrapper over Babel-Core. My Jekyll plugin should just leverage that when I
publish it.