Update draft;
This patch updates the curling post to be closer to a what my original
intention for it was. My original intention was to discuss different
`curl` commands I've used to debug CORS. While the history lesson is
good. Maybe I'll clean it up a bit more before publishing it later
today.