My big reveal is here. RIP iTerm2;
Pour one out for my longtime tool iTerm2. I have loved it and talked at
length about it. And, I would still recommend it to anyone starting to
use a Terminal emulator on macOS. But for me the time has come to
mature into more Unix philosophy tooling while also embracing
cross-platform tools. Alacritty, for now, is that tool for me. It's
focus is very reassuring and its dedication to speed is important to
me.